requirement.txt
astroid==1.6.5
autopep8==1.3.5
diff-match-patch==20121119
Django==2.0.7
django-ckeditor==5.4.0
django-crispy-forms==1.7.2
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.0.1
django-js-asset==1.0.0
django-reversion==2.0.13
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
isort==4.3.4
jdcal==1.4
jieba==0.39
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.12
odfpy==1.3.6
openpyxl==2.5.4
Pillow==5.1.0
pkg-resources==0.0.0
pycodestyle==2.4.0
pylint==1.9.2
pytz==2018.3
PyYAML==3.12
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
Whoosh==2.7.4
wrapt==1.10.11
xadmin==2.0.1
xlrd==1.1.0
xlwt==1.3.0
